+++ app/platforms/dc20/feat.rb
@@ -12,6 +12,7 @@ module Dc20
     SPELL_ORIGIN = 'spell'
     BASE_ANCESTRY_ORIGIN = 'base_ancestry'
     CLASS_ADDITIONAL_ORIGIN = 'class_additional'
+    WILD_FORM_ORIGIN = 'wild_form'
 
     STATIC = 'static' # рендерится текст
     UPDATE_RESULT = 'update_result' # рендерится, но затемняется
@@ -25,7 +26,7 @@ module Dc20
     COMBAT = 'combat'
     ONE_AT_SHORT_REST = 'one_at_short_rest' # 1 заряд восстанавливается при коротком отдыхе
 
-    SELECTABLE_ORIGINS = [0, 3, 4, 7].freeze
+    SELECTABLE_ORIGINS = [0, 3, 4, 7, 10].freeze
 
     enum :origin, {
       ANCESTRY_ORIGIN => 0,
@@ -37,7 +38,8 @@ module Dc20
       SUBCLASS_FLAVOR_ORIGIN => 6,
       SPELL_ORIGIN => 7,
       BASE_ANCESTRY_ORIGIN => 8,
-      CLASS_ADDITIONAL_ORIGIN => 9
+      CLASS_ADDITIONAL_ORIGIN => 9,
+      WILD_FORM_ORIGIN => 10
     }
     enum :kind, { STATIC => 0, UPDATE_RESULT => 1, HIDDEN => 2, ONE_FROM_LIST => 3, MANY_FROM_LIST => 4, TEXT => 5 }
     enum :limit_refresh, { SHORT_REST => 0, LONG_REST => 1, COMBAT => 2, ONE_AT_SHORT_REST => 3 }