+++ app/serializers/dc20/character_serializer.rb
@@ -12,13 +12,14 @@ module Dc20
                :stamina_points, :grit_points, :rest_points, :mana_spend_limit, :cantrips, :spells, :talent_points, :spell_attack,
                :selected_talents, :subclass, :speeds, :jump, :size, :breath, :spell_list, :spell_check, :spell_class,
                :selected_additional_talents, :ancestry_points, :damages, :resistances, :conditions_v2, :abilities, :martial_check,
-               :combat_expertise, :spell_filter
+               :combat_expertise, :spell_filter, :wild_form_available, :wild_form, :wild_forms
 
     delegate :features, :name, :modified_abilities, :health, :combat_mastery, :save_dc, :precision_defense, :breath, :size,
              :area_defense, :attack, :skills, :trades, :attribute_saves, :physical_save, :mental_save, :initiative, :attacks,
              :mana_points, :maneuver_points, :stamina_points, :grit_points, :rest_points, :damages, :mana_spend_limit,
              :stamina_spend_limit, :cantrips, :spells, :speeds, :jump, :visions, :spell_check, :spell_list, :spell_class,
-             :martial_check, :spell_attack, :combat_expertise, :spell_filter, to: :decorator
+             :martial_check, :spell_attack, :combat_expertise, :spell_filter, :wild_form_available, :wild_form, :wild_forms,
+             to: :decorator
     delegate :id, :created_at, :data, to: :object
     delegate :guide_step, :selected_talents, :subclass, :selected_additional_talents, :ancestry_points,
              :conditions_v2, :abilities, :attribute_points, :classes, :ancestries, :skill_points, :skill_expertise_points,