+++ db/migrate/20260809121536_add_index_to_types.rb
@@ -1,0 +1,8 @@
+class AddIndexToTypes < ActiveRecord::Migration[8.1]
+  disable_ddl_transaction!
+
+  def change
+    add_index :character_companions, :type, algorithm: :concurrently
+    add_index :characters, :type, algorithm: :concurrently
+  end
+end