+++ spec/controllers/frontend/dc20/characters/wild_forms_controller_spec.rb
@@ -1,0 +1,81 @@
+# frozen_string_literal: true
+
+describe Frontend::Dc20::Characters::WildFormsController do
+  let!(:user_session) { create :user_session }
+  let(:access_token) { Authkeeper::GenerateTokenService.new.call(user_session: user_session)[:result] }
+  let!(:user_character) { create :dc20_character, user: user_session.user }
+
+  describe 'GET#index' do
+    context 'for logged users' do
+      context 'for unexisting character' do
+        it 'returns error' do
+          get :index, params: { character_id: 'unexisting', charkeeper_access_token: access_token }
+
+          expect(response).to have_http_status :not_found
+        end
+      end
+
+      context 'for user character' do
+        context 'without wild_forms' do
+          it 'returns data', :aggregate_failures do
+            get :index, params: { character_id: user_character.id, charkeeper_access_token: access_token }
+
+            expect(response).to have_http_status :ok
+            expect(response.parsed_body['wild_forms'].size).to eq 0
+          end
+        end
+
+        context 'with wild_forms' do
+          before { create :dc20_wild_form, parent: user_character }
+
+          it 'returns data', :aggregate_failures do
+            get :index, params: { character_id: user_character.id, charkeeper_access_token: access_token }
+
+            expect(response).to have_http_status :ok
+            expect(response.parsed_body['wild_forms'].size).to eq 1
+          end
+        end
+      end
+    end
+  end
+
+  describe 'POST#create' do
+    context 'for logged users' do
+      context 'for unexisting character' do
+        it 'returns error' do
+          post :create, params: { character_id: 'unexisting', charkeeper_access_token: access_token }
+
+          expect(response).to have_http_status :not_found
+        end
+      end
+
+      context 'for user character' do
+        context 'with invalid params' do
+          let(:request) {
+            post :create, params: {
+              character_id: user_character.id, wild_form: { name: '' }, charkeeper_access_token: access_token
+            }
+          }
+
+          it 'returns error', :aggregate_failures do
+            expect { request }.not_to change(Dc20::WildForm, :count)
+            expect(response).to have_http_status :unprocessable_content
+          end
+        end
+
+        context 'with valid params' do
+          let(:request) {
+            post :create, params: {
+              character_id: user_character.id, wild_form: { name: 'Form' }, charkeeper_access_token: access_token
+            }
+          }
+
+          it 'creates wild form', :aggregate_failures do
+            expect { request }.to change(Dc20::WildForm, :count).by(1)
+            expect(response).to have_http_status :created
+          end
+        end
+      end
+    end
+  end
+end